Correcting issue where User was imported in notification views; Moving notification context_processor in notification app
