+++ notification/views/__init__.py
@@ -1,0 +1,4 @@
+from .notification import NotificationsListView, NotificationAdminView
+from .notification_settings import NotificationSettingEditView
+
+__all__ = [NotificationAdminView, NotificationsListView, NotificationSettingEditView]
