+++ notification/views/context_processors.py
@@ -1,0 +1,6 @@
+from django.conf import settings
+
+
+def daisy_notifications_enabled(request):
+    notifications_enabled = not getattr(settings, "NOTIFICATIONS_DISABLED", True)
+    return {"notifications_enabled": notifications_enabled}
